% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{augmentData}
\alias{augmentData}
\title{Augments the Master data.frame with additional information}
\usage{
augmentData(masterTsDataFrame)
}
\arguments{
\item{masterTsDataFrame}{The Master Time Series data.frame}
}
\value{
The Master data.frame with additinal data elements 
        weekday, weekend
}
\description{
Augments the Master data.frame with additional information
}

