% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{augmentIntraData}
\alias{augmentIntraData}
\title{Augments the intra day data.frame with additional information}
\usage{
augmentIntraData(inFrame)
}
\arguments{
\item{inFrame}{The Master Time Series data.frame}
}
\value{
The Master data.frame with additinal data elements 
        weekday, weekend, cum.sums of various variables
}
\description{
Augments the intra day data.frame with additional information
}

