% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{buildChartDay}
\alias{buildChartDay}
\title{Build Day timeseries Chart}
\usage{
buildChartDay(data, y.axes)
}
\arguments{
\item{data}{data.frame}

\item{y.axes}{Names of the Y-axes data, as a vector of characters}
}
\value{
A plot
}
\description{
Plots charts that have been selected as most relevant.
}

