% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{buildChartIntra}
\alias{buildChartIntra}
\title{Build Intraday Chart}
\usage{
buildChartIntra(data, y.axes)
}
\arguments{
\item{data}{data.frame}

\item{y.axes}{Names of the Y-axes data, as a vector of characters}
}
\value{
A plot
}
\description{
Plots intraday charts for the most relevant activities
}

