% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{connectToAPI}
\alias{connectToAPI}
\title{Connects to Fibit API}
\usage{
connectToAPI(appname, key, secret)
}
\arguments{
\item{appname}{Name of the Fitbit App}

\item{key}{Fitbit API Client key}

\item{secret}{Fibit API Client secret}
}
\value{
A Fitbit API token, that will be cached
}
\description{
Connects to the Fitbit API with OAuth 2. 
See https://dev.fitbit.com/docs/oauth2/
}

