% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{createDependentVariableFrame}
\alias{createDependentVariableFrame}
\title{Creates a data.frame with only goal variables}
\usage{
createDependentVariableFrame(master, goal)
}
\arguments{
\item{master}{Master data.frame}

\item{goal}{Goal variable}
}
\description{
Creates a data.frame with only goal variables
}

