% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{createGoalVariableVector}
\alias{createGoalVariableVector}
\title{Creates a vector of goal variables}
\usage{
createGoalVariableVector(master, goal)
}
\arguments{
\item{master}{Master data.frame}

\item{goal}{Goal variable}
}
\description{
Creates a vector of goal variables
}

