% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{createIntraFrame}
\alias{createIntraFrame}
\title{Creates the intraday Frame}
\usage{
createIntraFrame(folder)
}
\arguments{
\item{folder}{The folder in which JSON files will be read.}
}
\description{
Creates the intraday Frame
}

