% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{fetchAndStoreFile}
\alias{fetchAndStoreFile}
\title{A function to fetch data from fitbit.com and store the data as json files
to a user defined file destination}
\usage{
fetchAndStoreFile(request_url, token, targetFile)
}
\arguments{
\item{request_url}{the URL for GET call to fitbit.com}

\item{token}{the authorization token. Assumption that oauth2.0 has been done earlier in the flow}

\item{targetFile}{the place to store the json file}
}
\value{
the targetfile
}
\description{
A function to fetch data from fitbit.com and store the data as json files
to a user defined file destination
}

