% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{fetchIntraResourceData}
\alias{fetchIntraResourceData}
\title{Loads the JSON files for intraday data and returns a data.frame}
\usage{
fetchIntraResourceData(folder, resource, files)
}
\arguments{
\item{folder}{the folder to source the files from}

\item{resource}{the type of resource(Eg: calories, steps, distance etc)}

\item{files}{the list of files to look into for fetch}
}
\value{
Resource data.frame
}
\description{
Loads the JSON files for intraday data and returns a data.frame
}

