% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitAnalyzer.R
\docType{class}
\name{FitAnalyzer}
\alias{FitAnalyzer}
\title{R6 class for Analyzing Fitbit  Data}
\format{A \code{\link{R6Class}} generator object}
\usage{
FitAnalyzer
}
\description{
This is a R6 class for Analyzing Fitbit data.
Class is useful one you have the fitbit json files loaded in a folder as a starting point
}
\keyword{data}

