% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{getAPIScope}
\alias{getAPIScope}
\title{Get API scope}
\usage{
getAPIScope()
}
\value{
A vector of scope
}
\description{
Gets the scopes that will be retrieved by the API request to fitbit.
See https://dev.fitbit.com/docs/oauth2/#scope
}

