% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{getDailyResourcePathList}
\alias{getDailyResourcePathList}
\title{Returns a list of Fitbit Daily activities}
\usage{
getDailyResourcePathList()
}
\value{
A list
}
\description{
Returns a list of Fitbit Daily activities
}

