% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{getIntradayResourcePathList}
\alias{getIntradayResourcePathList}
\title{Returns a list of Fitbit Intraday activities}
\usage{
getIntradayResourcePathList()
}
\value{
A list
}
\description{
Returns a list of Fitbit Intraday activities
}

