% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{makeAPIRequest}
\alias{makeAPIRequest}
\title{Make API Request}
\usage{
makeAPIRequest(type, activity, start.date, end.date, api.token)
}
\arguments{
\item{type}{Type of time series. Must be 'day' or 'intraday'}

\item{activity}{Type of activity. See below for details.}

\item{start.date}{Start date in format YYYY-mm-dd}

\item{end.date}{End date in format YYYY-mm-dd}

\item{api.token}{API token for connection to Fitbit API}
}
\value{
The request response
}
\description{
Makes request to Fitbit API, and stores the response into a variable.
}

