% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{markValidRows}
\alias{markValidRows}
\title{A function that incorporates rules for marking if the data entry in MasterTSFrame are valid or not}
\usage{
markValidRows(masterTsDataFrame)
}
\arguments{
\item{masterTsDataFrame}{The Master Time Series data Frame}
}
\value{
The marked Master Data Frame. i.e column valid is added at the end of the data.frame
}
\description{
A function that incorporates rules for marking if the data entry in MasterTSFrame are valid or not
}

