% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{properCase}
\alias{properCase}
\title{Proper Case}
\usage{
properCase(x)
}
\arguments{
\item{x}{A string}
}
\value{
A string with proper case
}
\description{
Sets a string to proper case, i.e. upper case for the first letter of each word
}

