% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{writeToJSON}
\alias{writeToJSON}
\title{Write to JSON}
\usage{
writeToJSON(content, path, type, activity, start.date)
}
\arguments{
\item{content}{JSON content to be written to file}

\item{path}{Path to folder where files will be created}

\item{type}{Type of time series. Must be 'day' or 'intraday'}

\item{activity}{Type of activity. See below for details.}

\item{start.date}{Start date}
}
\description{
Writes API response content to JSON files, in a specific folder
}

