% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitUtil.R
\name{createTsMasterFrame}
\alias{createTsMasterFrame}
\title{Creates the Master data.frame from Timeseries JSON files.}
\usage{
createTsMasterFrame(tsFileFolder, resourcePath = getDailyResourcePathList())
}
\arguments{
\item{tsFileFolder}{Folder containing all time-series files. Naming convention for files is max-[resource].json}

\item{resourcePath}{the resource paths to look. Default will get getDailyResourcePathList()}
}
\value{
The Master data.frame
}
\description{
Creates the Master data.frame from Timeseries JSON files.
}

