% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitAnalyzer.R
\docType{class}
\name{FitAnalyzer}
\alias{FitAnalyzer}
\title{R6 class for Analyzing Fitbit  Data}
\format{A \code{\link{R6Class}} generator object}
\usage{
FitAnalyzer
}
\description{
FitAnalyzer is an R6 class for analyzing Fitbit data. It is an opinionated implementation of a particular workflow for analysis. 
For people attempting to conduct their own analysis in a different fashion you should use the more generic functions implemented in #' FitUtil.
The workflow implemented for FitAnalyzer is the following
1.	Create the FitAnalyzer with the goal variable for analysis. Eg: Calories or steps or distance. The goal variable is your personal #' goal that you are trying to analyze better.
2.	Call \code{findImportantVariables} to understand the most important variables unique to you that enable meeting your goal. 
3.	Call \code{showMostImportantCharts} to get relevant charts that are unique to your data
4.	Call \code{predictGoal} to get a prediction on performance of the goal
You can conduct two types of analysis based on the type of dataset in consideration. \code{analysis.type} can be 'intra.day'
analysis or it can be 'daily' .
}
\section{Methods}{

\describe{
  \item{ 
    \code{getAnalysisFrame(folder , analysis.type)}
    {This method uses \code{folder} \code{analysis.type} as an argument
    to return a data.frame that is clean and augmented with 
    additional features like weekend.}
  }
}
}
\keyword{data}

